cf-redirect-OIDC/deploy/requirements.txt
# Function dependencies, for example:
# package>=version
firebase-admin>=4.5.0
google-cloud-error-reporting>=1.1.0